third_party/WebKit/LayoutTests/http/tests/origin_trials/webexposed/webusb-origin-trial-interfaces-script-added.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>WebUSB - origin trial is enabled by script-added meta tag</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/origin-trials-helper.js"></script>
<script>
// Generate token with the command:
// generate_token.py  http://127.0.0.1:8000 WebUSB --expire-timestamp=2000000000
var old_trial_token = "Ak+Acw9/y+acA5hzmO0vOifI2G0n6g1+CiHBlyMOjjspFsPlTfvvI0nQBLCUlDkXkMQzSXzB97JmRaNNjpSyVQ0AAABOeyJvcmlnaW4iOiAiaHR0cDovLzEyNy4wLjAuMTo4MDAwIiwgImZlYXR1cmUiOiAiV2ViVVNCIiwgImV4cGlyeSI6IDIwMDAwMDAwMDB9";
// generate_token.py  http://127.0.0.1:8000 WebUSB2 --expire-timestamp=2000000000
var token = "AmgWWAL+v07ElzPwp9oXvY1ldvKTbMlEUFopboMutESil7WUfQNVdYFUzfai6GRq5MndTa+ln+w1gLlRw8VshwAAAABPeyJvcmlnaW4iOiAiaHR0cDovLzEyNy4wLjAuMTo4MDAwIiwgImZlYXR1cmUiOiAiV2ViVVNCMiIsICJleHBpcnkiOiAyMDAwMDAwMDAwfQ==";

test(t => {
  var interfaces =
      OriginTrialsHelper.get_interface_names(this,
        ['Navigator',
         'USBAlternateInterface',
         'USBConfiguration',
         'USBConnectionEvent',
         'USBDevice',
         'USBEndpoint',
         'USBInTransferResult',
         'USBInterface',
         'USBIsochronousInTransferPacket',
         'USBIsochronousInTransferResult',
         'USBIsochronousOutTransferPacket',
         'USBIsochronousOutTransferResult',
         'USBOutTransferResult'],
        {'Navigator': ['usb']});
  console.log('Interfaces before adding trial token\n' + interfaces);
}, "WebUSB related interfaces before adding trial token via script.");

OriginTrialsHelper.add_token(token);

test(t => {
  var interfaces =
      OriginTrialsHelper.get_interface_names(this,
        ['Navigator',
         'USBAlternateInterface',
         'USBConfiguration',
         'USBConnectionEvent',
         'USBDevice',
         'USBEndpoint',
         'USBInTransferResult',
         'USBInterface',
         'USBIsochronousInTransferPacket',
         'USBIsochronousInTransferResult',
         'USBIsochronousOutTransferPacket',
         'USBIsochronousOutTransferResult',
         'USBOutTransferResult'],
        {'Navigator': ['usb']});
  console.log('Interfaces after adding trial token\n' + interfaces);
}, "WebUSB related interfaces after adding trial token via script.");
</script>
